% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapping.R
\name{etrid2gr}
\alias{etrid2gr}
\title{From a transcript identifier to \code{GRanges} object}
\usage{
etrid2gr(etrid, ens)
}
\arguments{
\item{etrid}{A vector of Ensembl transcript identifiers.}

\item{ens}{A instance of class \code{Mart} from biomaRt. If
missing, \code{useMart("ensembl", "hsapiens_gene_ensembl")} is
used.}
}
\value{
A \code{Granges} object.
}
\description{
This function takes on or more Ensembl transcript identifiers,
queries Biomart and constructs a \code{GRanges} object as would
\code{Gviz::BiomartGeneRegionTrack} for a genomic region (in fact,
currently most of the code has been taken from
\code{Gviz::.fetchBMData} and \code{GViz::.chrName} is used to
validate chromosome names).
}
\examples{
id <- c("ENST00000612959", "ENST00000317091")
etrid2gr(id[1])
(gr <- etrid2gr(id))
## make it a GRangesList
(grl <- split(gr, mcols(gr)$transcript))
}
\author{
Laurent Gatto
}

