% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapToGenome}
\alias{mapToGenome}
\title{Map peptides on genomic coordinates}
\usage{
mapToGenome(pObj, grObj)
}
\arguments{
\item{pObj}{An \code{Proteins} instance of length 1, contain
\code{pfeatures} and \code{pranges}.}

\item{grObj}{A \code{GRanges} object containing the exon positions
of the protein in \code{pObj}, as created by
\code{\link{etrid2grl}}.}
}
\value{
A \code{Granges} object with the peptide coordinates on
the genome.
}
\description{
TODO
}
\details{
To calculate the genomic coordinates of peptides we

1) Calculate the contiguous exon coordinates on the cDNA sequence.
2) Calculate the peptide coordinates on the cDNA sequence from the
   original peptide ranges along the protein sequence.
3) We calculate the indices of the exons in which all the peptides
   start and end on the cDNA sequence.
4) We infer the coordinates on the genome from the actual peptide
   starting/ending position and exon index on the cDNA sequence, the
   exons coordinates on the cDNA sequences and the matching exons
   coordinates on the genome.
}
\author{
Laurent Gatto
}
\seealso{
The \code{\link{proteinCoding}} function to remove
non-protein coding ranges before maping peptides to their genomic
coordinates.
}

