\name{p}
\alias{p}
\docType{data}

\title{
  An \code{Proteins} object
}


\description{

  A small example \code{Proteins} test instance. This object is likely
  to change on a regular basis. It will be described more thoroughly
  when it becomes stable. 

}

\seealso{
  The \code{Pbase-data} vignette.
}

\usage{data(p)}

\examples{
data(p)
p
}

\keyword{datasets}
