% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotAsAnnotationTrack}
\alias{plotAsAnnotationTrack}
\alias{plotAsGeneRegionTrack}
\title{Plot gene region and annotation tracks}
\usage{
plotAsAnnotationTrack(x, ..., genome = "hg38", plot = TRUE)

plotAsGeneRegionTrack(..., genome = "hg38", plot = TRUE)
}
\arguments{
\item{x}{A \code{Granges} object containing peptides genomics
coordinates, typically generated by
\code{\link{pmapToGenome}}. These ranges are converted to a
\code{AnnotationTrack}.}

\item{...}{One or more \code{GRanges} instances, typically
resulting from calling \code{\link{etrid2grl}}, or, a single
\code{GRangesList}. These ranges are converted to
\code{GeneRegionTrack} instances.}

\item{genome}{A \code{character} of length 1, giving the name of
the genome. Default is \code{"hg38"}. If \code{NULL}, no
chromosome and axis tracks are displayed.}

\item{plot}{A \code{logical} defining if the figure should be
plotted. Default is \code{TRUE}.}
}
\value{
Used for its plotting side effects. Invisible returns a
list of tracks.
}
\description{
These functions convert ranges of peptides or exons to
\code{AnnotationTrack} or \code{GeneRegionTrack} objects from the
\code{Gviz} package and produces the corresponding plot. The
\code{genome} argument controls whether additional ideogram and
axis tracks are to be plotted. \code{plotAsAnnotationTrack} plots
peptides that span multiple exons in red and connects them with a
grey line. See \code{\link{pmapToGenome}} for example code.
}
\author{
Laurent Gatto
}

