% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges.R
\name{etrid2grl}
\alias{etrid2grl}
\title{From a transcript identifier to \code{GRanges} object}
\usage{
etrid2grl(etrid, ens, use.names = FALSE)
}
\arguments{
\item{etrid}{A vector of Ensembl transcript identifiers.}

\item{ens}{A instance of class \code{Mart} from biomaRt. If
missing, \code{useMart("ensembl", "hsapiens_gene_ensembl")} is
used.}

\item{use.names}{If set to \code{TRUE} and \code{etrid} has names,
then the latter are used to name the output.}
}
\value{
A \code{GRangesList} object of length
\code{length(etrid)}.
}
\description{
This function takes on or more Ensembl transcript identifiers,
queries Biomart and constructs a \code{GRangesList} object as
would \code{Gviz::BiomartGeneRegionTrack} for a genomic region (in
fact, currently most of the code has been taken from
\code{Gviz::.fetchBMData} and \code{GViz::.chrName} is used to
validate chromosome names).
}
\examples{
id <- c("ENST00000612959", "ENST00000317091")
grl1 <- etrid2grl(id[1])
grl1
grl <- etrid2grl(id)
stopifnot(all.equal(id, names(grl)))
}
\author{
Laurent Gatto
}
