% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isReverse}
\alias{isForward}
\alias{isReverse}
\title{Are all the ranges on the same strand}
\usage{
isReverse(gr)

isForward(gr)
}
\arguments{
\item{gr}{A \code{GRanges} object.}
}
\value{
A logical if \emph{all} the ranges in the \code{gr} object
are on the \code{"-"} (or \code{"+"} for code{isForward}) strand.
}
\description{
Checks if all ranges of a GRanges object are reverse.
}
\author{
Laurent Gatto
}

