% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapToGenome}
\alias{mapToGenome}
\title{Map peptides on genomic coordinates}
\usage{
mapToGenome(pObj, grObj)
}
\arguments{
\item{pObj}{An \code{Proteins} instance of length 1, contain
\code{pfeatures} and \code{pranges}.}

\item{grObj}{A \code{GRanges} object containing the exon positions
of the protein in \code{pObj}, as created by
\code{\link{etrid2gr}}.}
}
\value{
A \code{Granges} object with the peptide coordinates on
the genome.
}
\description{
TODO
}
\author{
Laurent Gatto
}

