\name{p}
\alias{p}
\alias{pms}
\docType{data}

\title{
  An \code{Proteins} object and its MSMS spectra
}


\description{

  A small example \code{Proteins} test instance. This object is likely
  to change on a regular basis. It will be described more thoroughly
  when it becomes stable. The MSMS spectra that were searched against
  the database are available in the \code{pms} \code{MSnExp} object.

}

\seealso{
  The \code{Pbase-data} vignette.
}

\usage{
data(p)
data(pms)
}

\examples{
data(p)
p
data(pms)
pms
}

\keyword{datasets}
